backend/requirements.txt
fastapi
httpx
Pillow
boto3
pydantic-settings
python-multipart
python-telegram-bot
PyYAML
sqlmodel
uvicorn[standard]
pytest

